pyproject.toml
[tool.poetry]
name = "eventbus-learning"
version = "0.1.0"
description = "Simple Eventbus project for learning purposes."
authors = ["Catriona Bennett <catrionamairi@blueyonder.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
aws-cdk-lib = "^2.161.1"
constructs = "^10.3.0"
python = "^3.12"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.10"
boto3 = "^1.35.25"
botocore = "^1.35.25"
cfn-lint = "^1.16.0"
coverage = "^7.6.1"
pre-commit = "^3.8.0"
pytest = "^8.3.3"
pytest-env = "^1.1.5"
ruff = "^0.6.7"
safety = "^3.2.7"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
